Move signer to new storage interface
This adds proto types for our state, and switches the signer to use
the new interface. This has been done in a way to avoid to much code
churn, rather than being optimally re-designed. We may want to revisit
how the rotation works with the new interface at some point