Support connectors with HTTP handlers
This is required for connectors that handle their own callbacks or have
other needs to serve HTTP.

Serving HTTP introduced yet another bit of configuration that will need
to be passed from the server to the connector: the URL and/or path where
the `Connector`'s `http.Handler` will be mounted. In our case we've
chosen `/connector/:id`, but it's obviously not something we want
connectors hard-coding. Connectors may also not have an awareness of
their own ID. It's also possible that Deci itself is mounted in a path
(e.g., `/deci`), so the full path may end up being something like
`/deci/connector/:id`. Most connectors will need to know this URL to
properly generate links or HTTP redirects in its own `http.Handler`.

My first tactic to support additional configuration options was to
create a higher level `ConnectorBuilder` function that was passed a set
of `ConnectorOptions` and returned `(Connector, error`). Something like:

```go
type ConnectorBuilder func(opts ConnectorOpts) (Connector, error)

connectorBulders := map[string]ConnectorBuilder{
  "static": func(opts ConnectorOpts) (Connector, error) {
    // opts.Authenticator would make the Authenticator available
    // opts.BaseURL would make the BaseURL available
    //
    // Additional fields could be added to opts in a
    // backwards-compatible way later

    // ...
    return connector, nil
   },
}

server, err := NewServer(..., connectorBuilders, ...)
```

While I think this would solve the problem, it adds significant
complexity to the constructor creation process. It also makes connectors
more awkward to test on their own because a `ConnectorBuilder` must be
used (which looks awkward in a test that otherwise doesn't care how the
`Connector` is built, because it's being tested independently of
`Server`) or the `ConnectorBuilder` is left untested, which leaves
important code untested.

I settled instead on constructing a `Server` first, then adding
`Connector` instances later. The `Server`'s public API becomes
usable when building the `Connector`, but smaller interfaces (e.g.,
`Authenticator` or a `string` from the return value of `ConnectorURL`)
can be passed as constructor arguments or set directly on `Connector`
structs. This makes it more pleasant to test a `Connector` on its own
(without `Server`) and avoids additional types like `ConnectorBuilder`
and such that look awkward in Go code.

I've also ripped out the awareness of `URL.Path` in `Server`'s mux. This
follows the convention of other Go `http.Handler` (e.g., `FileServer`)
which always serve relative to their own view of `/`, but can be mounted
at any path as long as `http.StripPrefix` is used. This reduces the need to
test non-`/` paths and allows more flexiblility in mounting the `Server`
in `main`.