Ensure signed JWT's have KeyID Header
We were seeing issues where the signed JWT's did not have a KeyID header set to
the right value, which cause issues locating the corresponding public key to
verify the signature against.

Turns out this is because we weren't passing the correct value to the
SigningKey - it takes an `interface{}`, which worked for both the raw RSA key
which has no concept of a JWK KeyID, as well as a `jose.JSONWebKey` which does.
Fix by passing the JWK to the SigningKey, and update the test to ensure all
signed tokens contain a KeyID.