Bring back connector.Initialize
Previously reverted in 2f875bc9af260fa6d20481f276bb8dcaee59d22b, I'm
proposing bringing it back because certain design patterns around
chaining connectors is _much_ easier if the chained connector can
intercept (and possibly wrap) the `Authenticator` as it is passed
through at initialization time. While in general I support Go structs
that are fully initialized when constructed, `Initialize` is, in this
case, the more flexible design pattern.

If we ever find a better design pattern for this, I'd be happy to
revisit it.