Token generation and management
We want to be able to issue opaque tokens that we can persist in a safe manner,
so that data leaks do not contain usable data.

This is implemented by generating tokens as an ID.TOKEN part. The ID is always
clear, and can be used to locate items in the database. The token part is
always stored using bcrypt.